requirements.txt
crewai==0.28.8
crewai-tools==0.1.5
python-dotenv
PyYAML==6.0.1
setuptools